backend/requirements.txt
# ArtinSmartRealty V2 Backend Dependencies
# Python 3.11+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Redis (Context Management)
redis==5.0.1

# Telegram Bot
python-telegram-bot==21.0.1

# AI / LLM
google-generativeai==0.8.3  # Gemini Vision for smart property extraction (FREE!)

# PDF Generation & Processing
reportlab==4.0.9
Pillow==10.2.0
PyPDF2==3.0.1
pytesseract==0.3.13  # OCR for image text extraction

# Excel Export
openpyxl==3.1.2
pandas==2.1.4

# Scheduling
apscheduler==3.10.4

# Audio Processing (Voice Messages)
pydub==0.25.1

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# HTTP Client
httpx==0.27.0
aiohttp==3.9.1

# Authentication
PyJWT==2.8.0

# Utilities
python-dotenv==1.0.0
pytz==2024.1
